Split MMA and HelpCentre page skeletons
