+++ app/client/components/MMAPageSkeleton.tsx
@@ -1,5 +1,3 @@
-import { css } from "@emotion/core";
-import { space } from "@guardian/src-foundations";
 import { Location } from "@reach/router";
 import React from "react";
 import {
@@ -9,8 +7,6 @@ import {
 } from "../../shared/productTypes";
 import { MenuSpecificNavItem, NAV_LINKS } from "./nav/navConfig";
 import { PageContainer, WithStandardTopMargin } from "./page";
-import { SectionHeader } from "./sectionHeader";
-import { SectionPageContainer } from "./sectionPageContainer";
 import { Spinner } from "./spinner";
 
 interface LocationObject {
@@ -113,49 +109,13 @@ const MMALocationObjectArr: LocationObject[] = [
   }
 ];
 
-const nonMMALocationObjectArr: LocationObject[] = [
-  {
-    title: "Help Centre",
-    path: "/help-centre",
-    selectedNavItem: NAV_LINKS.help
-  },
-  {
-    title: "Need to contact us about something?",
-    path: "/contact-us",
-    selectedNavItem: NAV_LINKS.help
-  }
-];
-
-const PageSkeleton = () => {
+const MMAPageSkeleton = () => {
   return (
     <Location>
       {({ location }) => {
         const selectedMMALocationObject = MMALocationObjectArr.filter(
           currentObject => location.pathname === currentObject.path
         )[0];
-        const selectedNonMMALocationObject = nonMMALocationObjectArr.filter(
-          currentObject => location.pathname.startsWith(currentObject.path)
-        )[0];
-
-        if (selectedNonMMALocationObject) {
-          return (
-            <>
-              <SectionHeader title={selectedNonMMALocationObject.title} />
-              <SectionPageContainer sectionTitle="&nbsp;">
-                <div
-                  css={css`
-                    margin-bottom: ${space[24]}px;
-                  `}
-                >
-                  <WithStandardTopMargin>
-                    <Spinner />
-                  </WithStandardTopMargin>
-                  <div style={{ height: "50vh" }} />
-                </div>
-              </SectionPageContainer>
-            </>
-          );
-        }
 
         if (selectedMMALocationObject) {
           return (
@@ -174,4 +134,4 @@ const PageSkeleton = () => {
   );
 };
 
-export default PageSkeleton;
+export default MMAPageSkeleton;
